* Add a bunch of LUA- and other missing files to VS8 project file. * Disabled 64 bit compatibility warnings and disabled   comparison between signed and unsigned integer warnings.   (fixing them just makes code uglier because of all the (int) casts) * Fixed all other warnings in VS8 except those in hpiutil2, streflop, lua.   (it's mostly int->boolean performance warnings have been fixed) * Fixed FTBFS when DIRECT_CONTROL_ALLOWED wasn't #defined. * Made it use CrashHandler.cpp on MSVC too (as opposed to CrashReporter).
git-svn-id: https://spring.clan-sy.com/svn/spring/trunk@3775 37977431-3df6-0310-b722-df95706aa16b

+++ rts/Lua/LuaVFS.cpp
@@ -74,7 +74,7 @@ LuaVFS::AccessMode LuaVFS::GetMode(lua_State* L, int index, bool synced)
 	if (!lua_isnumber(L, index)) {
 		luaL_error(L, "Bad VFS access mode");
 	}
-	const AccessMode mode = (AccessMode)lua_tonumber(L, index);
+	const AccessMode mode = (AccessMode)(int)lua_tonumber(L, index);
 	if ((mode < 0) || (mode > LAST_MODE)) {
 		luaL_error(L, "Bad VFS access mode");
 	}
